% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fields.R
\name{all_fields}
\alias{all_fields}
\title{All Study Fields}
\usage{
all_fields()
}
\value{
vector
}
\description{
Return all available study fields on clinicaltrials.gov
}
\examples{
\dontrun{
fields <- all_fields()
}

}
