% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_fields.R
\name{default_fields}
\alias{default_fields}
\title{Default fields}
\usage{
default_fields()
}
\value{
vector
}
\description{
Return selected default fields to use when query clinicaltrials.gov
}
\examples{
\dontrun{
fields <- default_fields()
}

}
