% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_by_id.R
\name{search_by_id}
\alias{search_by_id}
\title{search using STUDYID or NCT num}
\usage{
search_by_id(
  id,
  area = c("OrgStudyId", "SecondaryId"),
  coverage = c("Contains", "FullMatch", "StartsWith", "EndsWith"),
  fields = default_fields(),
  min_rnk = 1,
  max_rnk = 1000,
  fmt = "csv",
  output = "WIDE"
)
}
\arguments{
\item{id}{study id or NCT num}

\item{area}{declares which search area/fields should be searched}

\item{coverage}{degree to which a search term needs to match the text in an API field. There are four choices: Contains(default), FullMatch, StartsWith, EndsWith}

\item{fields}{fields to be returned (defaulted to default_fields())}

\item{min_rnk}{minimum returned searches}

\item{max_rnk}{maximum returned searches (note: this cannot be > 1000)}

\item{fmt}{format as per API, defaulted to csv}

\item{output}{can be WIDE or LONG format}
}
\value{
data.frame
}
\description{
search clinicaltrials.gov using any study id or NCT number
}
\examples{
\dontrun{
res <- search_by_id("BP28248")
res <- search_by_id("BP28248", coverage = "FullMatch")
res <- search_by_id("NCT01874691", area = c("NCTId"), fields = c("OrgStudyId", "SecondaryId"))
res <- search_by_id("BP28248", fields = c("NCTId"))
}

}
