% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ctgov.R
\name{search_ctgov}
\alias{search_ctgov}
\title{Search clinicaltrials.gov}
\usage{
search_ctgov(
  expr,
  fields = default_fields(),
  min_rnk = 1,
  max_rnk = 1000,
  fmt = "csv",
  output = "WIDE"
)
}
\arguments{
\item{expr}{expression as defined by clinicaltrials.gov (https://clinicaltrials.gov/api/gui/ref/syntax)}

\item{fields}{fields to be returned (defaulted to default_fields())}

\item{min_rnk}{minimum returned searches}

\item{max_rnk}{maximum returned searches (note: this cannot be > 1000)}

\item{fmt}{format as per API, defaulted to csv}

\item{output}{can be WIDE or LONG format}
}
\value{
data.frame
}
\description{
search clinicaltrials.gov using the Study Fields API (https://clinicaltrials.gov/api/gui/demo/simple_study_fields).
Search can be specified with expression and the fields required.
}
\examples{
\dontrun{
res <- search_ctgov("heart+attack", max_rnk = 10)
res <- search_ctgov("BP28248")
res <- search_ctgov("BP28248", fields = all_fields()) #note: all_fields() takes a long time to run
res <- search_ctgov("heart+attack", fields = c("OrgStudyId","SecondaryId","NCTId"))

}
}
